requirements.txt
alembic==1.13.3                                                                                                                                                               
annotated-types==0.7.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.3
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.0.7
flask-openapi3==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.5
MarkupSafe==2.1.2
mysqlclient==2.2.4
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4